Refactor positional labels via the new properties
