+++ quarkdown-core/src/main/kotlin/com/quarkdown/core/ast/quarkdown/block/Figure.kt
@@ -2,9 +2,9 @@ package com.quarkdown.core.ast.quarkdown.block
 
 import com.quarkdown.core.ast.Node
 import com.quarkdown.core.ast.SingleChildNestableNode
-import com.quarkdown.core.ast.attributes.CaptionableNode
 import com.quarkdown.core.ast.attributes.location.LocationTrackableNode
 import com.quarkdown.core.ast.base.inline.Image
+import com.quarkdown.core.ast.quarkdown.CaptionableNode
 import com.quarkdown.core.visitor.node.NodeVisitor
 
 /**
